Update index.html
trying code on saturday night.  The neigbhors arguing got my mind out of speed chess and interuppted my attention to the meidastouch live saturday evening episode of legal analysis regarding the former president of the united states that no one in my orbit of people want to talk about, but instead maintain a facade of denial of how to properly live in society.  One more than the other... it brings to mind a tweet, - 'the MAGA guy who doesn't know he is a MAGA guy, but holds MAGA values, is making too much noise by walking in his apartment causing the floors to creek, among other unintentional noises that the MAGA guy who lives underneath him, who happens to hail from the west side of Fort Worth Texas, an area known for its conservative and quiet control by large landowners or oil companies to become annoyed and seek retribution.  This evolved into a shouting match in the breeze way that startled me out of a pleasent Sunday afternoon buzz and in a spate of nervousness smoked a cig, pulled the shades, and walked around with piqued curiosity of how far the arguement would go.  It lasted a good 5 five minutes culminating in the loud deep voice of a man saying, "WHO THE FUCK DO YOU THINK YOU ARE?", and then a door slamming.  

Personally I think we should all have a group hug and bbq and talk about the latest Greg Abbott border controls that involve drowning 4 year olds, or the news I heard today of over 500 immigrants (i think mostly dead), in Mexico that died in shipping containers, but that may be kind of a touchy subject with these guys.

So I don't know why I typed this all here, but if anyone ever reads it, this is Larry Schnellmanns Sunday afternoon in Dallas wanting to be hanging out with my brother or my sisters in aluethra which is some exoctic carribbeen island.

Larry Schnellmann   7-22-2023